general: Fix potential nullpointer problem in destructors.
This is a rather tricky issue. In PHP objects are not stored in
variables, only pointers to the objects are. Using pass by value,
the object will still be the same, but the pointer will be a different
one. Using pass by reference, also the pointer will be the same.
By setting the pointer NULL, all other users of the pointer will get
a NULL value trying to access the object.