Parsers: Fix lazy initialization of nullable request values
PHP doesn't have an easy way to check whether a readonly property has been initialized to NULL.
There's multiple workarounds:

1) Reflection

You could use ReflectionProperty::isInitialized() to check whether the property has been
set or not. This works, but comes with the drawback of needing Reflection, which adds a
performance penalty on the code (however miniscule it might be).

This is also quite repetitive code, so we'd need to centralize it in an isInitialized()
method of an abstract RequestValueParser class that all parsers that need it would extend
from.

2) Alternative "not found" value

You could adopt a different value to indicate that a request value has not been found.
None of the existing scalar types qualify, however, since they could conflict with
legitimate usecases for those values, which leaves objects.

Such an object wouldn't hold any values, it would just be an empty shell.

This would also require to change semantics in the Request class to return such an object
instead of NULL for "not found" cases, which introduces quite a compatibility break with
the existing code, for seemingly no practical reason.

3) Track initialization state in an array

For parsers that parse a lot of request values (currently none exist, however), having
one property per value to track the initialization state may be cumbersome, but feels
like the easiest option to ensure correctness.

Array properties can't be fully readonly, meaning the initialization state could still be
modified at any point.

4) One property per value to track initialization state

Which leaves the least ugly version of one property to track the initialization state per
value. That property is defined to only be able to exist in a boolean state. It's either
TRUE or uninitialized, which makes it easy and performant to work with, just annoying.